[BE][refactor] 상품,브랜드,옵션,장바구니 기능 예외처리

+++ src/main/java/com/doubleowner/revibe/domain/cart/service/CartService.java
@@ -7,6 +7,8 @@ import com.doubleowner.revibe.domain.cart.repository.CartRepository;
 import com.doubleowner.revibe.domain.item.entity.Item;
 import com.doubleowner.revibe.domain.item.repository.ItemRepository;
 import com.doubleowner.revibe.domain.user.entity.User;
+import com.doubleowner.revibe.global.exception.CommonException;
+import com.doubleowner.revibe.global.exception.errorCode.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,6 +29,11 @@ public class CartService {
 
         Item item = itemRepository.findByIdOrElseThrow(requestDto.getItemId());
 
+        // 이미 장바구니에 해당상품이 존재 할 경우 예외처리
+        if(cartRepository.existsByUserIdAndItemId(loginUser.getId(),requestDto.getItemId())){
+            throw new CommonException(ErrorCode.ALREADY_EXIST,"이미 장바구니에 등록한 상품입니다.");
+        }
+
         Cart cart = new Cart(loginUser, item);
 
         cartRepository.save(cart);
@@ -46,6 +53,9 @@ public class CartService {
     @Transactional
     public void deleteCart(User loginUser, Long cartId) {
         Cart cart = cartRepository.findByIdAndUser(cartId, loginUser);
+        if(cart == null){
+            throw new CommonException(ErrorCode.NOT_FOUND_VALUE,"해당 장바구니 상품을 찾을 수 없습니다.");
+        }
         cartRepository.delete(cart);
     }
